+++ Kurnik/SQL_Editor.cpp
@@ -1,6 +1,9 @@
 #include "pch.h"
 #include "SQL_Editor.h"
-#include "Collected_Eggs_Model.h"
+#include "SQL table data models/Collected_Eggs_Model.h"
+#include "SQL table data models/Sold_Eggs_Model.h"
+#include "SQL table data models/Buyers_Model.h"
+#include "SQL table data models/Bought_Resources_Model.h"
 
 SQL_Editor::SQL_Editor(wxWindow* parent, SQLite::Database& database)
 	:wxPanel(parent), database(database)
@@ -11,7 +14,10 @@ SQL_Editor::SQL_Editor(wxWindow* parent, SQLite::Database& database)
 		wxDV_ROW_LINES | wxDV_VERT_RULES | wxDV_HORIZ_RULES);
 	data_view->SetMinSize(wxSize(400, 400));
 
-	active_data_model = data_models[0];
+	//active_data_model = data_models["Collected Eggs"];
+	active_data_model = data_models["Sold Eggs"];
+	//active_data_model = data_models["Buyers"];
+	//active_data_model = data_models["Bought Resources"];
 	data_view->AssociateModel(active_data_model);
 	for (const auto& col : active_data_model->getColumns()) {
 		data_view->AppendColumn(col);
@@ -23,9 +29,26 @@ SQL_Editor::SQL_Editor(wxWindow* parent, SQLite::Database& database)
 
 void SQL_Editor::init_data_models()
 {
-	data_models.reserve(1);
-
-	data_models.emplace_back(new Collected_Eggs_Model("Collected Eggs", "Collected Eggs", database));
+	{
+		std::string_view table{ "Collected Eggs" };
+		data_models.emplace(std::make_pair(table,
+			new Collected_Eggs_Model(table, table, database)));
+	}
+	{
+		std::string_view table{ "Sold Eggs" };
+		data_models.emplace(std::make_pair(table,
+			new Collected_Eggs_Model(table, "Sold Eggs + Buyers", database)));
+	}
+	{
+		std::string_view table{ "Buyers" };
+		data_models.emplace(std::make_pair(table,
+			new Collected_Eggs_Model(table, table, database)));
+	}
+	{
+		std::string_view table{ "Bought Resources" };
+		data_models.emplace(std::make_pair(table,
+			new Collected_Eggs_Model(table, table, database)));
+	}
 }
 
 void SQL_Editor::recreate_table()
