+++ Kurnik/pch.h
@@ -19,5 +19,6 @@
 #include <string>
 #include <string_view>
 #include <array>
+#include <map>
 #include <vector>
 #include <memory>
